# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fratelanza Management Console API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  password: string;
}

export interface AuthResult {
  success: boolean;
  /** @nullable */
  token?: string | null;
}

export interface DashboardSummary {
  totalRevenue: number;
  totalPaid: number;
  totalRemaining: number;
  totalNetProfit: number;
  totalExpenses: number;
  activeProjects: number;
  completedProjects: number;
  totalClients: number;
  totalFreelancers: number;
}

export interface ProfitByType {
  type: string;
  netProfit: number;
  count: number;
}

export interface PaymentAlert {
  id: number;
  projectName: string;
  clientName: string;
  remaining: number;
  nextPaymentDate: string;
}

export interface Project {
  id: number;
  type: string;
  projectName: string;
  /** @nullable */
  clientName?: string | null;
  clientPrice: number;
  totalCost: number;
  netProfit: number;
  /** @nullable */
  freelancerName?: string | null;
  freelancerCommission?: number;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  deadline?: string | null;
  status: string;
  paidAmount?: number;
  remainingAmount?: number;
  /** @nullable */
  nextPaymentDate?: string | null;
  /** @nullable */
  notes?: string | null;
  date: string;
}

export interface TeamMember {
  id: number;
  projectId: number;
  freelancerName: string;
  commission: number;
}

export type ProjectDetail = Project & {
  team?: TeamMember[];
};

export interface TeamMemberInput {
  freelancerName: string;
  commission: number;
}

export interface ProjectInput {
  type: string;
  projectName: string;
  clientName?: string;
  clientPrice: number;
  totalCost?: number;
  netProfit?: number;
  freelancerName?: string;
  freelancerCommission?: number;
  startDate?: string;
  deadline?: string;
  status?: string;
  paidAmount?: number;
  remainingAmount?: number;
  nextPaymentDate?: string;
  notes?: string;
  team?: TeamMemberInput[];
}

export interface ProjectUpdate {
  projectName?: string;
  clientName?: string;
  clientPrice?: number;
  totalCost?: number;
  netProfit?: number;
  freelancerCommission?: number;
  startDate?: string;
  deadline?: string;
  status?: string;
  paidAmount?: number;
  remainingAmount?: number;
  nextPaymentDate?: string;
  notes?: string;
  team?: TeamMemberInput[];
}

export interface PaymentInput {
  amount: number;
  nextPaymentDate?: string;
}

export interface Freelancer {
  code: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  spec?: string | null;
  /** @nullable */
  position?: string | null;
  earned: number;
  balance: number;
  rating: number;
}

export interface FreelancerInput {
  name: string;
  phone?: string;
  spec?: string;
  position?: string;
  earned?: number;
  balance?: number;
}

export interface FreelancerUpdate {
  name?: string;
  phone?: string;
  spec?: string;
  position?: string;
  earned?: number;
  balance?: number;
  rating?: number;
}

export interface Client {
  id: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  activity?: string | null;
  /** @nullable */
  project?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface ClientInput {
  name: string;
  phone?: string;
  address?: string;
  activity?: string;
  project?: string;
  notes?: string;
}

export interface ClientUpdate {
  name?: string;
  phone?: string;
  address?: string;
  activity?: string;
  project?: string;
  notes?: string;
}

export type ClientProfile = Client & {
  projects?: Project[];
  totalProjects?: number;
  totalValue?: number;
  totalPaid?: number;
  totalRemaining?: number;
};

export interface Template {
  id: number;
  category: string;
  name: string;
  cost: number;
  expenses?: number;
  multiplier?: number;
  broker?: number;
  students?: number;
}

export interface TemplateInput {
  category: string;
  name: string;
  cost: number;
  expenses?: number;
  multiplier?: number;
  broker?: number;
  students?: number;
}

export interface TemplateUpdate {
  category?: string;
  name?: string;
  cost?: number;
  expenses?: number;
  multiplier?: number;
  broker?: number;
  students?: number;
}

export interface Quote {
  id: number;
  clientName: string;
  /** @nullable */
  projectName?: string | null;
  price: number;
  /** @nullable */
  language?: string | null;
  date: string;
  /** @nullable */
  paymentTerms?: string | null;
  /** @nullable */
  milestones?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface QuoteInput {
  clientName: string;
  projectName?: string;
  price: number;
  language?: string;
  date?: string;
  paymentTerms?: string;
  milestones?: string;
  notes?: string;
}

export interface QuoteUpdate {
  clientName?: string;
  projectName?: string;
  price?: number;
  language?: string;
  paymentTerms?: string;
  milestones?: string;
  notes?: string;
}

export interface Expense {
  id: number;
  description: string;
  amount: number;
  date: string;
}

export interface ExpenseInput {
  description: string;
  amount: number;
  date?: string;
}

export interface ExpenseSummary {
  totalExpenses: number;
  count: number;
}

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: string;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectName?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}

export interface TaskInput {
  title: string;
  description?: string;
  status?: string;
  priority?: string;
  projectId?: number;
  assignedTo?: string;
  dueDate?: string;
}

export interface TaskUpdate {
  title?: string;
  description?: string;
  status?: string;
  priority?: string;
  projectId?: number;
  assignedTo?: string;
  dueDate?: string;
}

export interface FinanceReport {
  totalRevenue: number;
  totalPaid: number;
  totalRemaining: number;
  totalCost: number;
  totalNetProfit: number;
  totalExpenses: number;
  netBalance: number;
  projects: Project[];
}

export type ListProjectsParams = {
type?: string;
status?: string;
search?: string;
};

export type ListFreelancersParams = {
search?: string;
spec?: string;
};

export type ListClientsParams = {
search?: string;
project?: string;
};

export type ListTemplatesParams = {
category?: string;
};

export type ListQuotesParams = {
client?: string;
search?: string;
};

export type ListExpensesParams = {
startDate?: string;
endDate?: string;
};

export type GetExpenseSummaryParams = {
startDate?: string;
endDate?: string;
};

export type ListTasksParams = {
status?: string;
projectId?: number;
};

export type GetFinanceReportParams = {
startDate?: string;
endDate?: string;
};