# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fratelanza Management Console API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with master password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid password
  /auth/verify:
    post:
      operationId: verifyPassword
      tags: [auth]
      summary: Verify master password for sensitive operations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Password verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid password

  # ── Dashboard ─────────────────────────────────────────────────────────
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients with optional search
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: project
          schema:
            type: string
      responses:
        "200":
          description: Clients list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Add a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client 360 profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientProfile"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Templates (Services) ──────────────────────────────────────────────
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List general expenses
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
        - in: query
          name: endDate
          schema:
            type: string
      responses:
        "200":
          description: Expenses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Add a general expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/summary:
    get:
      operationId: getExpenseSummary
      tags: [expenses]
      summary: Get expense totals and summary
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
        - in: query
          name: endDate
          schema:
            type: string
      responses:
        "200":
          description: Expense summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseSummary"
  /expenses/{id}:
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete a general expense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Tasks ─────────────────────────────────────────────────────────────
...
  /finance/report:
    get:
      operationId: getFinanceReport
      tags: [expenses]
      summary: Get full P&L financial report
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
        - in: query
          name: endDate
          schema:
            type: string
      responses:
        "200":
          description: Finance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceReport"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
    AuthResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        token:
          type: ["string", "null"]

    # ── Dashboard ──────────────────────────────────────────────────────
...
    Project:
      type: object
      required: [id, type, projectName, clientPrice, totalCost, netProfit, status, date]
      properties:
        id:
          type: integer
        type:
          type: string
        projectName:
          type: string
        clientName:
          type: ["string", "null"]
        clientPrice:
          type: number
        totalCost:
          type: number
        netProfit:
          type: number
        freelancerName:
          type: ["string", "null"]
        freelancerCommission:
          type: number
        startDate:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        status:
          type: string
        paidAmount:
          type: number
        remainingAmount:
          type: number
        nextPaymentDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        date:
          type: string
...
    Client:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        activity:
          type: ["string", "null"]
        project:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ClientInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        activity:
          type: string
        project:
          type: string
        notes:
          type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        activity:
          type: string
        project:
          type: string
        notes:
          type: string
    ClientProfile:
      allOf:
        - $ref: "#/components/schemas/Client"
        - type: object
          properties:
            projects:
              type: array
              items:
                $ref: "#/components/schemas/Project"
            totalProjects:
              type: integer
            totalValue:
              type: number
            totalPaid:
              type: number
            totalRemaining:
              type: number

    # ── Template ───────────────────────────────────────────────────────
...
    Expense:
      type: object
      required: [id, description, amount, date]
      properties:
        id:
          type: integer
        description:
          type: string
        amount:
          type: number
        date:
          type: string
    ExpenseInput:
      type: object
      required: [description, amount]
      properties:
        description:
          type: string
        amount:
          type: number
        date:
          type: string
    ExpenseSummary:
      type: object
      required: [totalExpenses, count]
      properties:
        totalExpenses:
          type: number
        count:
          type: integer

    # ── Task ───────────────────────────────────────────────────────────
...
    FinanceReport:
      type: object
      required: [totalRevenue, totalPaid, totalRemaining, totalCost, totalNetProfit, totalExpenses, netBalance, projects]
      properties:
        totalRevenue:
          type: number
        totalPaid:
          type: number
        totalRemaining:
          type: number
        totalCost:
          type: number
        totalNetProfit:
          type: number
        totalExpenses:
          type: number
        netBalance:
          type: number
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"